pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "code-review-benchmark"
version = "0.1.0"
description = "Benchmark suite for open source AI code review tools"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12,<1",
    "pydantic>=2,<3",
    "pyyaml>=6,<7",
    "gitpython>=3.1,<4",
    "openai>=1.30,<2",
    "rich>=13,<14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "ruff>=0.4,<1",
]

[project.scripts]
crb = "code_review_benchmark.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/code_review_benchmark"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
